requirements.txt
pillow==12.2.0
python-dotenv==1.2.2
requests==2.33.1
ollama==0.6.1